Move sassbiotope model to bims
