+++ bims/models/__init__.py
@@ -24,3 +24,4 @@ from bims.models.taxon_group import *  # noqa
 from bims.models.vernacular_name import *  # noqa
 from bims.models.river_catchment import *  # noqa
 from bims.models.fbis_uuid import *  # noqa
+from bims.models.sass_biotope import *  # noqa
