+++ sass/models/__init__.py
@@ -1,6 +1,5 @@
 from sass.models.river import *  # noqa
 from sass.models.site_visit import *  # noqa
-from sass.models.sass_biotope import *  # noqa
 from sass.models.rate import *  # noqa
 from sass.models.sass_biotope_fraction import *  # noqa
 from sass.models.sass_taxon import *  # noqa
